архитектурные правки, контракт для money_service

+++ tegtory/domain/interfaces/user.py
@@ -3,7 +3,7 @@ from tegtory.domain.interfaces.base import CrudRepository
 
 
 class UserRepository(CrudRepository[User]):
-    async def subtract(self, user_id: int, amount: int) -> None:
+    async def subtract(self, user_id: int, amount: float) -> None:
         pass
 
     async def add(self, user_id: int, amount: int) -> None:
